Initial commit.
There are two tests, testBoard is closer to how the first t-boards will look. The basic mechanics for moving letters around and having respawners are all working, but it only supports touch input. Multitouch is fully handled, so it should work well for cooperative play. Next will be finishing testBoard, then working on a tool to define boards.